Revert enumeration from model back
