+++ powerauth-fido2/src/main/java/com/wultra/powerauth/fido2/rest/model/enumeration/AttestationType.java
@@ -1,6 +1,6 @@
 /*
  * PowerAuth Server and related software components
- * Copyright (C) 2023 Wultra s.r.o.
+ * Copyright (C) 2024 Wultra s.r.o.
  *
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU Affero General Public License as published
@@ -14,16 +14,20 @@
  *
  * You should have received a copy of the GNU Affero General Public License
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
+ *
  */
 
-package com.wultra.security.powerauth.fido2.model.enumeration;
+package com.wultra.powerauth.fido2.rest.model.enumeration;
 
 /**
- * Elliptic curve type enumeration.
+ * Attestation type.
  *
- * @author Petr Dvorak, petr@wultra.com
+ * @author Roman Strobl, roman.strobl@wultra.com
  */
-public enum CurveType {
-    P256,
-    UNKNOWN
+public enum AttestationType {
+
+    BASIC,
+    SELF,
+    NONE
+
 }
