+++ powerauth-fido2/src/main/java/com/wultra/powerauth/fido2/rest/model/enumeration/ECKeyType.java
@@ -1,6 +1,6 @@
 /*
  * PowerAuth Server and related software components
- * Copyright (C) 2024 Wultra s.r.o.
+ * Copyright (C) 2023 Wultra s.r.o.
  *
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU Affero General Public License as published
@@ -14,20 +14,18 @@
  *
  * You should have received a copy of the GNU Affero General Public License
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
- *
  */
 
-package com.wultra.security.powerauth.fido2.model.enumeration;
+package com.wultra.powerauth.fido2.rest.model.enumeration;
 
 /**
- * FIDO2 configuration key strings.
+ * Elliptic key type enumeration.
  *
  * @author Roman Strobl, roman.strobl@wultra.com
  */
-public class Fido2ConfigKeys {
+public enum ECKeyType {
 
-    public static final String CONFIG_KEY_ALLOWED_ATTESTATION_FMT = "fido2_attestation_fmt_allowed";
-    public static final String CONFIG_KEY_ALLOWED_AAGUIDS = "fido2_aaguids_allowed";
-    public static final String CONFIG_KEY_ROOT_CA_CERTS = "fido2_root_ca_certs";
+    COMPRESSED,
+    UNCOMPRESSED
 
 }
