+++ powerauth-fido2/src/main/java/com/wultra/powerauth/fido2/rest/model/enumeration/Fmt.java
@@ -16,14 +16,32 @@
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-package com.wultra.security.powerauth.fido2.model.enumeration;
+package com.wultra.powerauth.fido2.rest.model.enumeration;
+
+import java.util.List;
 
 /**
- * Signature algorithm enumeration.
+ * Attestation format enumeration.
  *
  * @author Petr Dvorak, petr@wultra.com
  */
-public enum SignatureAlgorithm {
-    ES256,
-    UNKNOWN
+public enum Fmt {
+
+    FMT_PACKED("packed"),
+    FMT_NONE("none");
+
+    private final String value;
+
+    public static final List<String> allowedFmt = List.of(
+            FMT_PACKED.value, FMT_NONE.value
+            // not supported: "tpm", "android-key", "android-safetynet", "apple", "fido-u2f"
+    );
+
+    Fmt(String value) {
+        this.value = value;
+    }
+
+    public String getValue() {
+        return value;
+    }
 }
