+++ powerauth-fido2/src/main/java/com/wultra/powerauth/fido2/rest/model/validator/RegistrationRequestValidator.java
@@ -21,10 +21,10 @@ package com.wultra.powerauth.fido2.rest.model.validator;
 import com.wultra.powerauth.fido2.rest.model.entity.*;
 import com.wultra.powerauth.fido2.rest.model.request.RegistrationRequestWrapper;
 import com.wultra.security.powerauth.fido2.model.entity.AuthenticatorParameters;
-import com.wultra.security.powerauth.fido2.model.enumeration.CurveType;
-import com.wultra.security.powerauth.fido2.model.enumeration.ECKeyType;
-import com.wultra.security.powerauth.fido2.model.enumeration.Fmt;
-import com.wultra.security.powerauth.fido2.model.enumeration.SignatureAlgorithm;
+import com.wultra.powerauth.fido2.rest.model.enumeration.CurveType;
+import com.wultra.powerauth.fido2.rest.model.enumeration.ECKeyType;
+import com.wultra.powerauth.fido2.rest.model.enumeration.Fmt;
+import com.wultra.powerauth.fido2.rest.model.enumeration.SignatureAlgorithm;
 import com.wultra.security.powerauth.fido2.model.request.RegistrationRequest;
 import com.wultra.security.powerauth.fido2.model.response.AuthenticatorAttestationResponse;
 import io.getlime.security.powerauth.crypto.lib.util.Hash;
